requirements.txt
prefect>=2.14.0
pandas>=2.0.0
pyyaml>=6.0
python-slugify>=8.0.0
numpy>=1.24.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0